+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - update deptrac in composer to use new deptrac/deptrac package
 
+### Fixed
+- CategoryIdFilter didn't implemented matches method from FilterInterface
+
 ## [4.1.0] - 2025-06-11
 This is stable release for v4.1.0. No changes have been made since v4.1.0-rc.1.